// client/stocklinx app/src/redux/customField/constant.ts
export enum customFieldConst {
  FETCH_CUSTOMFIELDS_REQUEST = "FETCH_CUSTOMFIELDS_REQUEST",
  FETCH_CUSTOMFIELDS_SUCCESS = "FETCH_CUSTOMFIELDS_SUCCESS",
  FETCH_CUSTOMFIELDS_FAILURE = "FETCH_CUSTOMFIELDS_FAILURE",
  FETCH_CUSTOMFIELD_REQUEST = "FETCH_CUSTOMFIELD_REQUEST",
  FETCH_CUSTOMFIELD_SUCCESS = "FETCH_CUSTOMFIELD_SUCCESS",
  FETCH_CUSTOMFIELD_FAILURE = "FETCH_CUSTOMFIELD_FAILURE",
  FETCH_CUSTOMFIELD_COUNTS_REQUEST = "FETCH_CUSTOMFIELD_COUNTS_REQUEST",
  FETCH_CUSTOMFIELD_COUNTS_SUCCESS = "FETCH_CUSTOMFIELD_COUNTS_SUCCESS",
  FETCH_CUSTOMFIELD_COUNTS_FAILURE = "FETCH_CUSTOMFIELD_COUNTS_FAILURE",
  CREATE_CUSTOMFIELD_REQUEST = "CREATE_CUSTOMFIELD_REQUEST",
  CREATE_CUSTOMFIELD_SUCCESS = "CREATE_CUSTOMFIELD_SUCCESS",
  CREATE_CUSTOMFIELD_FAILURE = "CREATE_CUSTOMFIELD_FAILURE",
  CREATE_RANGE_CUSTOMFIELD_REQUEST = "CREATE_RANGE_CUSTOMFIELD_REQUEST",
  CREATE_RANGE_CUSTOMFIELD_SUCCESS = "CREATE_RANGE_CUSTOMFIELD_SUCCESS",
  CREATE_RANGE_CUSTOMFIELD_FAILURE = "CREATE_RANGE_CUSTOMFIELD_FAILURE",
  UPDATE_CUSTOMFIELD_REQUEST = "UPDATE_CUSTOMFIELD_REQUEST",
  UPDATE_CUSTOMFIELD_SUCCESS = "UPDATE_CUSTOMFIELD_SUCCESS",
  UPDATE_CUSTOMFIELD_FAILURE = "UPDATE_CUSTOMFIELD_FAILURE",
  REMOVE_CUSTOMFIELD_REQUEST = "REMOVE_CUSTOMFIELD_REQUEST",
  REMOVE_CUSTOMFIELD_SUCCESS = "REMOVE_CUSTOMFIELD_SUCCESS",
  REMOVE_CUSTOMFIELD_FAILURE = "REMOVE_CUSTOMFIELD_FAILURE",
  REMOVE_RANGE_CUSTOMFIELD_REQUEST = "REMOVE_RANGE_CUSTOMFIELD_REQUEST",
  REMOVE_RANGE_CUSTOMFIELD_SUCCESS = "REMOVE_RANGE_CUSTOMFIELD_SUCCESS",
  REMOVE_RANGE_CUSTOMFIELD_FAILURE = "REMOVE_RANGE_CUSTOMFIELD_FAILURE",
  SET_CUSTOMFIELD = "SET_CUSTOMFIELD",
  CLEAR_CUSTOMFIELD = "CLEAR_CUSTOMFIELD",
  SET_CUSTOMFIELDS = "SET_CUSTOMFIELDS",
  CLEAR_CUSTOMFIELDS = "CLEAR_CUSTOMFIELDS",
}